Time improvements (#50)
* Add native Datetime struct
* Add PartialOrd and Ord derives
* Implement formatting for LocalDate, LocalTime, LocalDatetime, Datetime (matches formatting of corresponding types in chrono)
  LocalDate uses `yyyy-MM-dd`
  LocalTime uses `HH:mm:ss`, `HH:mm:ss.fff` or `HH:mm:ss.ffffff` depending on the number of fractional digits
  LocalDatetime uses `{date}T{time}` for Debug and `{date} {time}` for Display
  Datetime uses `{date}T{time}Z` for Debug and `{date} {time} UTC` for Display
* add to_micros accessor functions
* Define MAX/MIN constants for types and enforce in the constructor
* fix bug in LocalDateTime conversion which didn't account for the difference between unix and postgres epoch
* fix bug in LocalDateTime conversion which didn't handle div/rem on negative values correctly

+++ edgedb-protocol/src/codec.rs
@@ -832,7 +832,7 @@ impl Codec for LocalTime {
             _ => Err(errors::invalid_value(type_name::<Self>(), val))?,
         };
         buf.reserve(8);
-        buf.put_i64(val.micros);
+        buf.put_i64(val.micros as i64);
         Ok(())
     }
 }
